Move URL and QueryParams to new '_urls.py' module
